Change user.test to index.test, mimic folder structure
